+++ gimpformats/GimpPrecision.py
@@ -4,7 +4,7 @@
 # from .GimpImageHierarchy import GimpImageHierarchy
 from __future__ import annotations
 
-from binaryiotools import IO
+from .binaryiotools import IO
 
 
 class Precision: