+++ gimpformats/GimpVbrBrush.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 """Pure python implementation of the gimp vbr brush format.
 """
 from __future__ import annotations