+++ gimpformats/GimpVectors.py
@@ -1,11 +1,9 @@
-#!/usr/bin/env python3
 """Stuff related to vectors/paths within a gimp document.
 """
 # from .GimpIOBase import GimpIOBase
 from __future__ import annotations
 
-from binaryiotools import IO
-
+from .binaryiotools import IO
 from .GimpParasites import GimpParasite