+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2023
+
+- `binaryiotools` back in project
+
 ## 2022.0.1 - 2022/01/28
 
 - Using pytest